+++ src/Capital_Good.hpp
@@ -13,7 +13,7 @@ using namespace std;
 class Capital_Good: public General_Good{
 
     public:
-    Capital_Good(Capital_Firm_Agent *  seller_pointer, float seller_price, int seller_quantity, int life_span);
+    Capital_Good(Capital_Firm_Agent *  seller_pointer, float seller_price, long long seller_quantity, int life_span);
     Capital_Good(const Capital_Good& other);  // Copy constructor
 
     // Depreciation
