+++ src/Consumer_Firm_Agent.cpp
@@ -154,9 +154,9 @@ void Consumer_Firm_Agent::Determine_New_Production(){
     bool emission_high = emission_overshoot > 0; 
 
     // Calculate how many offsets you need to buy to reach the average unit emissions
-    int emission_offsets_needed = 0;
+    long long emission_offsets_needed = 0;
     // Add the amount we need to buy so that our inventory is at average emissions
-    emission_offsets_needed += static_cast<int>(emission_overshoot * inventory);
+    emission_offsets_needed += static_cast<long long>(emission_overshoot * inventory);
 
 
     good_price_past = good_price_current; // store current price incase we want to see the change
@@ -189,7 +189,7 @@ void Consumer_Firm_Agent::Determine_New_Production(){
     else if (!price_high && inventory_high && emission_high) {
         production_planned*= (1.0-prod_change/2.0);
         // Add the amount we need to buy to offset the emissions of the goods we plan to produce
-        emission_offsets_needed += static_cast<int>(production_planned * unit_emissions_adj);
+        emission_offsets_needed += static_cast<long long>(production_planned * unit_emissions_adj);
         Buy_Emission_Offsets(emission_offsets_needed);
         
     } // Case 7: Inventory high, price high, emissions low -> Decrease production slightly + decrease price slightly
@@ -227,9 +227,9 @@ void Consumer_Firm_Agent::Determine_New_Production(){
 
 /* Function to buy the required amount of emission offsets, and seek a long term loan if necessary
 */
-void Consumer_Firm_Agent::Buy_Emission_Offsets(int n_offsets) {
+void Consumer_Firm_Agent::Buy_Emission_Offsets(long long n_offsets) {
         // Calculate total cost of emissions we wish to buy 
-        int emission_offset_needed_cost = n_offsets * pPublic_Info_Board->Get_Emission_Offset_Price();
+        long long emission_offset_needed_cost = n_offsets * pPublic_Info_Board->Get_Emission_Offset_Price();
         long_term_funding_gap =  max(emission_offset_needed_cost - cash_on_hand,static_cast<long long>(0)); // Funding gap must be positive
         if (long_term_funding_gap > 0){
             Seek_Long_Term_Loan();
